+++ tests/testthat/test-ce_plot.R
@@ -1,17 +1,4 @@
 test_that("ce_plot works", {
-  get_nmb_training <- get_nmb_sampler(
-    qalys_lost = function() rnorm(1, 0.33, 0.03),
-    wtp = 28000,
-    high_risk_group_treatment_effect = function() exp(rnorm(n = 1, mean = log(0.58), sd = 0.43)),
-    high_risk_group_treatment_cost = function() rnorm(n = 1, mean = 161, sd = 49),
-    use_expected_values = TRUE
-  )
-  get_nmb_evaluation <- get_nmb_sampler(
-    qalys_lost = function() rnorm(1, 0.33, 0.03),
-    wtp = 28000,
-    high_risk_group_treatment_effect = function() exp(rnorm(n = 1, mean = log(0.58), sd = 0.43)),
-    high_risk_group_treatment_cost = function() rnorm(n = 1, mean = 161, sd = 49)
-  )
   sim_obj <- readRDS(test_path("fixtures", "predictNMBsim_ce_object.rds"))
 
   expect_error(ce_plot(sim_obj), "'ref_col'")
@@ -38,6 +25,18 @@ test_that("ce_plot works", {
   sim_ce_plot <- ce_plot(sim_obj, ref_col = "none", shape = 2)
   vdiffr::expect_doppelganger(
     "ce_plot.predictNMBsim(obj) triangles",
+    sim_ce_plot
+  )
+
+  sim_ce_plot <- ce_plot(sim_obj, ref_col = "none", shape = "cost-effective")
+  vdiffr::expect_doppelganger(
+    "ce_plot.predictNMBsim(obj) ce-shapes",
+    sim_ce_plot
+  )
+
+  sim_ce_plot <- ce_plot(sim_obj, ref_col = "none", shape = "method")
+  vdiffr::expect_doppelganger(
+    "ce_plot.predictNMBsim(obj) method-shapes",
     sim_ce_plot
   )
 
@@ -72,12 +71,6 @@ test_that("ce_plot works", {
     )
   )
 
-  get_nmb_evaluation <- get_nmb_sampler(
-    qalys_lost = function() rnorm(1, 0.33, 0.03),
-    wtp = function() rnorm(1, 28000, 1000),
-    high_risk_group_treatment_effect = function() exp(rnorm(n = 1, mean = log(0.58), sd = 0.43)),
-    high_risk_group_treatment_cost = function() rnorm(n = 1, mean = 161, sd = 49)
-  )
 
   sim_obj <- readRDS(test_path("fixtures", "predictNMBsim_object.rds"))
   expect_error(
@@ -96,6 +89,13 @@ test_that("ce_plot works", {
   sim_obj <- readRDS(test_path("fixtures", "predictNMBsim_ce_uncertain-wtp_object.rds"))
   expect_message(
     ce_plot(sim_obj, ref_col = "none"),
+    "10000 samples"
+  )
+
+  sim_obj <- readRDS(test_path("fixtures", "predictNMBsim_ce_uncertain-wtp_object.rds"))
+
+  expect_message(
+    ce_plot(sim_obj, ref_col = "none", shape = "method"),
     "10000 samples"
   )