+++ NEWS.md
@@ -1,7 +1,17 @@
 # predictNMB (development version)
 
-* add `wtp_linetype` argument to `ce_plot()` (and default to `"dashed"`) to 
+* `ce_plot()`
+
+  * add `wtp_linetype` argument to `ce_plot()` (and default to `"dashed"`) to 
   given control to user and differentiate C-E plane from main axes.
+  
+  * add `shape` argument which maps to the shape aesthetic in `geom_point()`. 
+  User can also specify `shape = "cost-effective"` to map it to whether that
+  point is under the WTP/cost-effectiveness plane, or `shape = "method"` to map
+  it to the cutpoint method.
+  
+  * add `add_prop_ce` argument. If `TRUE`, the % of points for that
+  cutpoint method is concatenated to the the method name in the figure legend.
 
 # predictNMB 0.2.0